wavelet: extract pure DWT + analyzer modules (library factor-out)
Pull the DWT math (dwt.js) and per-frame feature computation (waveletAnalyzer.js)
out of the worker into pure, import-anywhere modules. The worker is now a thin
transport shell around createWaveletAnalyzer — so the headless harness runs the
EXACT same code that runs live. This is the start of the standalone-library split.
